Tile layer for decoding improved. cache removed... Looking for another solution

+++ src/plugins/plugin-mapbox-gl/custom-layers/bitmap-layer/index.js
@@ -73,7 +73,7 @@ export default class BitmapLayer extends Layer {
 
     this.setState({
       numInstances: 1,
-      positions: new Float64Array(12)
+      positions: new Float32Array(12)
     });
   }
